+++ src/LaravelCacheAdapter.php
@@ -43,7 +43,7 @@ class LaravelCacheAdapter implements CacheInterface
      */
     public function get($key)
     {
-        $this->getCache()->get($this->generateKey($key));
+        return $this->getCache()->get($this->generateKey($key));
     }
 
     /**
@@ -76,13 +76,18 @@ class LaravelCacheAdapter implements CacheInterface
     /**
      * The AWS CacheInterface takes input in seconds, but the Laravel Cache classes use minutes. To support
      * this intelligently, we round up to one minute for any value less than 60 seconds, and round down to
-     * the nearest whole minute for any value over one minute.
+     * the nearest whole minute for any value over one minute. First, if the passed in TTL is 0 we return
+     * 0 to allow an unlimited cache lifetime.
      *
      * @param $ttl
      * @return float|int
      */
     protected function convertTtl($ttl)
     {
+        if ($ttl == 0) {
+            return 0;
+        }
+
         $minutes = floor($ttl / 60);
 
         if ($minutes == 0) {